Fix a bunch of bugs in the functions and a lot more stuff

+++ definicoes.h
@@ -23,7 +23,7 @@ typedef struct {
 } NEIGHBORHOOD;
 
 /////////////////////////////////////////////////////// STRUCTURES
-typedef float PARTICLE[MAX_VARIABLES];
+typedef double PARTICLE[MAX_VARIABLES];
 
 typedef struct {
 	int x;
